requirements.txt
annotated-types==0.7.0
babel==2.17.0
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
docxcompose==1.4.0
docxtpl==0.20.1
et_xmlfile==2.0.0
fonttools==4.59.2
Jinja2==3.1.6
kiwisolver==1.4.9
loguru==0.7.3
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
win32_setctime==1.2.0
